requirements.txt
httpx
python-telegram-bot
python-dotenv
aiofiles
